+++ components/header-content.tsx
