+++ components/main-content.tsx
