+++ components/styles/audio-player.style.ts
