+++ components/styles/index.ts
