+++ components/types/api/artist-response.interface.ts
