+++ components/types/api/favorite-response.interface.ts
