+++ components/types/artists.interface.ts
@@ -1,4 +1,4 @@
-import { ArtistResponse } from "./responses/artist-response.interface";
+import { ArtistResponse } from "./api/artist-response.interface";
 
 export interface ArtistsProps {
   artists: Array<ArtistResponse>;
