+++ components/types/custom/index.ts
@@ -1,0 +1,3 @@
+export * from "./tabs.enum";
+export * from "./song-data.interface";
+export * from "./song-fn.interface";
