+++ components/types/custom/song-data.interface.ts
