+++ components/types/custom/song-fn.interface.ts
@@ -3,5 +3,4 @@ import { SongData } from "./song-data.interface";
 export interface SongFn {
   getSongPath: () => SongData;
   setSongPath: (path: string, id: string, visible: boolean) => void;
-  openSearchDialog: () => void;
 }
