+++ components/types/song.interface.ts
@@ -1,5 +1,5 @@
-import { SongResponse } from "./responses/song-response.interface";
-import { SongFn } from "./top-nav.interface";
+import { SongResponse } from "./api/song-response.interface";
+import { SongFn } from "./custom/song-fn.interface";
 
 export interface SongProps {
   songFn: SongFn;
